feat(home): 小津移到右下角，意见反馈浮球并入小津气泡
用户指定：小津默认放右下角；原先占着右下的「意见反馈」FloatButton 去掉，
入口吸收进小津。

- 位置：.xiaojin-pet 左右翻转（right:24 / 移动端 right:14），气泡右对齐、
  尾巴移到右侧；「不再显示」✕ 从右上挪到左上，避免探出屏幕右缘。
- 反馈：FeedbackButton 整个删除（Layout 里 isHome 渲染点一并移除），其中的
  Modal 抽成受控的 FeedbackModal 复用；小津气泡底部新增「意见反馈」文字入口，
  仍只对登录用户显示（submitFeedback 需要登录态，与原浮球同门槛）。Modal 挂在
  气泡之外——点入口时气泡随即关闭，弹窗不能跟着卸载。

实测（真浏览器）：右下定位/气泡右对齐/尾巴指向、匿名不见入口、伪造登录态后
入口可见且弹窗完整可用、390px 与 360px 气泡不出界。CDP 里 JS click 触发的
Modal 会卡在 opacity:0（rc-motion 的 rAF 被节流，已知 CDP 假象），真实点击
路径正常。

测试：XiaojinPet 新增两条（登录可见反馈入口并开弹窗 / 匿名不可见），突变
实测均红（去掉 user 门槛、去掉 setFeedbackOpen 各挂一条）。全量 vitest 本机
仍有既有的 admin 页超时抖动（单独跑 18/18 过，母分支基线同样挂），以 CI 为准。
