+++ frontend/src/pages/HomePage.tsx
@@ -352,8 +352,8 @@ export default function HomePage() {
 
       </section>
 
-      {/* 左下角的小津。position:fixed，不参与 hero 的定高布局；
-          右下角留给 Layout 里的 FeedbackButton。 */}
+      {/* 右下角的小津。position:fixed，不参与 hero 的定高布局。
+          原先占右下角的意见反馈浮球已并入小津气泡底部。 */}
       <XiaojinPet />
     </div>
   );
